Changed SK24Light into a util file
SK24Light has been switched from a subsystem into a utility class that is simpler to use. It still works with the subsystem json, now requiring an init() method if the json value is true. To use in general code simply add SKCANLight as a parameter and run the specified methods without worrying about the subsystem json.